+++ lib/flkit.dart
@@ -2,3 +2,4 @@
 library;
 
 export 'commands/create_command.dart';
+export 'flkit_cli.dart';